[All] Fix remained warnings.

+++ lib/host/wasi/wasifunc.cpp
@@ -924,7 +924,7 @@ Expect<uint32_t> WasiFdPread::body(Runtime::Instance::MemoryInstance *MemInst,
   /// Store read bytes length.
   *NRead = preadv(Entry->second.HostFd, SysIOVS, IOVSLen, Offset);
 
-  if (unlikely(*NRead < 0)) {
+  if (unlikely(static_cast<ssize_t>(*NRead) < 0)) {
     return convertErrNo(errno);
   }
 
@@ -1046,7 +1046,7 @@ Expect<uint32_t> WasiFdPwrite::body(Runtime::Instance::MemoryInstance *MemInst,
 
   *NWritten = pwritev(Entry->second.HostFd, SysIOVS, IOVSLen, Offset);
 
-  if (unlikely(*NWritten < 0)) {
+  if (unlikely(static_cast<ssize_t>(*NWritten) < 0)) {
     return convertErrNo(errno);
   }
 
@@ -1111,7 +1111,7 @@ Expect<uint32_t> WasiFdRead::body(Runtime::Instance::MemoryInstance *MemInst,
   /// Store read bytes length.
   *NRead = readv(Entry->second.HostFd, SysIOVS, IOVSLen);
 
-  if (unlikely(*NRead < 0)) {
+  if (unlikely(static_cast<ssize_t>(*NRead) < 0)) {
     return convertErrNo(errno);
   }
 
@@ -1320,7 +1320,7 @@ Expect<uint32_t> WasiFdTell::body(Runtime::Instance::MemoryInstance *MemInst,
 
   /// Do lseek.
   *Offset = lseek(Entry->second.HostFd, 0, SEEK_CUR);
-  if (unlikely(*Offset < 0)) {
+  if (unlikely(static_cast<off_t>(*Offset) < 0)) {
     return convertErrNo(errno);
   }
 
@@ -1385,7 +1385,7 @@ Expect<uint32_t> WasiFdWrite::body(Runtime::Instance::MemoryInstance *MemInst,
 
   *NWritten = writev(Entry->second.HostFd, SysIOVS, IOVSLen);
 
-  if (unlikely(*NWritten < 0)) {
+  if (unlikely(static_cast<ssize_t>(*NWritten) < 0)) {
     return convertErrNo(errno);
   }
 
@@ -2316,7 +2316,7 @@ Expect<uint32_t> WasiSockRecv::body(Runtime::Instance::MemoryInstance *MemInst,
     *RoFlags |= __WASI_SOCK_RECV_DATA_TRUNCATED;
   }
 
-  if (unlikely(*RoDataLen < 0)) {
+  if (unlikely(static_cast<ssize_t>(*RoDataLen) < 0)) {
     return convertErrNo(errno);
   }
 
@@ -2391,7 +2391,7 @@ Expect<uint32_t> WasiSockSend::body(Runtime::Instance::MemoryInstance *MemInst,
   /// Store send bytes length and flags.
   *SoDataLen = sendmsg(Entry->second.HostFd, &SysMsgHdr, 0);
 
-  if (unlikely(*SoDataLen < 0)) {
+  if (unlikely(static_cast<ssize_t>(*SoDataLen) < 0)) {
     return convertErrNo(errno);
   }
